refactor: begin distinguishing between products and items
It didn't make much sense to have the term "keys" because not all
digital products will be a key. Instead, we changed it to item and
then changed item to product.

Moving forward, a product will be the term encompassing what is
being sold. An item will be the term for what is being delivered
as part of the product.