refactor: accommodate backend payment changes
The checkout process was changed to reflect the backend payment
changes. There is now a processing page the user will be taken to
while we are awaiting the webhook.

The crypto modal was reintroduced since BTCPay was added as a
payment provider. The workflow for BTCPay is different than it
was for CoinPayments, so most of the methods have been changed
or deleted.